Added back in a x10 correction factor for compressor knee.

+++ src/process.c
@@ -11,7 +11,7 @@
  *  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
  *  GNU General Public License for more details.
  *
- *  $Id: process.c,v 1.50 2004/04/11 15:30:10 jdepner Exp $
+ *  $Id: process.c,v 1.51 2004/04/11 23:24:58 theno23 Exp $
  */
 
 #include <math.h>
@@ -286,7 +286,7 @@ void run_eq(unsigned int port, unsigned int in_ptr)
 	peak = peak_data[i] * peak_data[i] + peak_data[BINS - i] *
 		peak_data[BINS - i];
 	if (peak > bin_peak[i]) {
-	    bin_peak[i] = sqrt(peak);
+	    bin_peak[i] = sqrtf(peak);
 	}
     }
     fftwf_execute(plan_cr);
